Guard delete tree to the owner only

+++ backend/src/Controller/TreeController.php
@@ -11,6 +11,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Http\Attribute\CurrentUser;
 
 #[Route('/tree')]
@@ -58,8 +59,12 @@ class TreeController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_tree_delete', methods: ['DELETE'])]
-    public function delete(Tree $tree, TreeRepository $treeRepository): Response
+    public function delete(Tree $tree, TreeRepository $treeRepository, #[CurrentUser] User $user): Response
     {
+        if ($tree->getOwner()->getUserIdentifier() !== $user->getUserIdentifier()) {
+            throw new AccessDeniedException();
+        }
+
         $treeRepository->remove($tree, true);
         return $this->json($tree);
     }
